fix: Remove updating type of genre
There was an edge case that if updating a game with 1 genre, dev or
platform and updating that 1 genre, dev or platform to another type
leaves the game without genres, developers or platforms. So we want to
have all the categories be like they are, and if the user wants to add
something to a game they can add the category first and then update the
game.